fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,20 +6,20 @@
 int main(void)
 {
 	int i = 0;
-	int x = 1;
-	int y = 2;
+	long x = 1;
+	long y = 2;
 
 	while (i < 50)
 	{
 	if (i == 0)
-	printf("%d", x);
+	printf("%ld", x);
 	else if (i == 1)
-	printf(", %d", y);
+	printf(", %ld", y);
 	else
 	{
 	y = y + x;
 	x = y - x;
-	printf(", %d", y);
+	printf(", %ld", y);
 	}
 	++i;
 	}
